pyproject.toml
[project]
name = "stream-agent"
version = "0.1.0"
description = "Megaforce Social Media API - Twitter/X integration with approval workflow"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "arcadepy",
    "python-dotenv",
    "langchain",
    "langchain-core",
    "langchain-openai",
    "langchain-anthropic",
    "arcade-ai>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "email-validator>=2.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.10.0",
    "psycopg2-binary>=2.9.0",
    "langchain-google-genai>=2.1.0",
]

[project.scripts]
api = "api.main:app"

[tool.hatch.build.targets.wheel]
packages = ["stream_agent"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
]
